+++ .idea/libraries/Maven__art_uniroma2_it_org_openrdf_sesame_sesame_onejar_2_7_10.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: art.uniroma2.it.org.openrdf.sesame:sesame-onejar:2.7.10">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/art/uniroma2/it/org/openrdf/sesame/sesame-onejar/2.7.10/sesame-onejar-2.7.10.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/art/uniroma2/it/org/openrdf/sesame/sesame-onejar/2.7.10/sesame-onejar-2.7.10-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/art/uniroma2/it/org/openrdf/sesame/sesame-onejar/2.7.10/sesame-onejar-2.7.10-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
