+++ .idea/libraries/Maven__au_com_bytecode_opencsv_2_4.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: au.com.bytecode:opencsv:2.4">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/au/com/bytecode/opencsv/2.4/opencsv-2.4.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/au/com/bytecode/opencsv/2.4/opencsv-2.4-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/au/com/bytecode/opencsv/2.4/opencsv-2.4-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
