+++ .idea/libraries/Maven__com_github_rvesse_airline_io_2_8_5.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: com.github.rvesse:airline-io:2.8.5">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/github/rvesse/airline-io/2.8.5/airline-io-2.8.5.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/com/github/rvesse/airline-io/2.8.5/airline-io-2.8.5-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/github/rvesse/airline-io/2.8.5/airline-io-2.8.5-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
