+++ .idea/libraries/Maven__jakarta_inject_jakarta_inject_api_1_0_3.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: jakarta.inject:jakarta.inject-api:1.0.3">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/jakarta/inject/jakarta.inject-api/1.0.3/jakarta.inject-api-1.0.3.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/jakarta/inject/jakarta.inject-api/1.0.3/jakarta.inject-api-1.0.3-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/jakarta/inject/jakarta.inject-api/1.0.3/jakarta.inject-api-1.0.3-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
