+++ .idea/libraries/Maven__org_slf4j_slf4j_jdk14_1_7_36.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.slf4j:slf4j-jdk14:1.7.36">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/slf4j/slf4j-jdk14/1.7.36/slf4j-jdk14-1.7.36.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/slf4j/slf4j-jdk14/1.7.36/slf4j-jdk14-1.7.36-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/slf4j/slf4j-jdk14/1.7.36/slf4j-jdk14-1.7.36-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
