+++ src/main/java/com/csv2rdf/csv/RowNumberProvider.java
@@ -1,0 +1,9 @@
+package com.csv2rdf.csv;
+
+
+public class RowNumberProvider extends ValueProvider {
+
+    protected String provideValue(int rowIndex, String[] row) {
+        return String.valueOf(rowIndex);
+    }
+}
