+++ src/main/java/com/csv2rdf/csv/RowValueProvider.java
@@ -1,0 +1,15 @@
+package com.csv2rdf.csv;
+
+import java.nio.charset.Charset;
+
+public class RowValueProvider extends ValueProvider {
+    private final int colIndex;
+
+    RowValueProvider(int colIndex) {
+        this.colIndex = colIndex;
+    }
+
+    protected String provideValue(int rowIndex, String[] row) {
+        return row[colIndex];
+    }
+}
