+++ src/main/java/com/csv2rdf/csv/StatementGenerator.java
@@ -1,0 +1,22 @@
+package com.csv2rdf.csv;
+
+import org.openrdf.model.*;
+
+public class StatementGenerator {
+    private final ValueGenerator<Resource> subject;
+    private final ValueGenerator<URI> predicate;
+    private final ValueGenerator<Value> object;
+
+    StatementGenerator(ValueGenerator<Resource> s, ValueGenerator<URI> p, ValueGenerator<Value> o) {
+        this.subject = s;
+        this.predicate = p;
+        this.object = o;
+    }
+
+    Statement generate(int rowIndex, String[] row) {
+        Resource s = subject.generate(rowIndex, row);
+        URI p = predicate.generate(rowIndex, row);
+        Value o = object.generate(rowIndex, row);
+        return CSV2RDF.FACTORY.createStatement(s, p, o);
+    }
+}
