+++ src/main/java/com/csv2rdf/csv/ValueGenerator.java
@@ -1,0 +1,7 @@
+package com.csv2rdf.csv;
+
+import org.openrdf.model.Value;
+
+interface ValueGenerator<V extends Value> {
+    V generate(int rowIndex, String[] row);
+}
